edit last task

+++ python_introduction/finance_calculator.py
@@ -30,7 +30,7 @@ monthly_income = int(input('Enter your monthly income: '))
 monthly_expenses = int(input('Enter your total monthly expenses: '))
 monthly_saving = monthly_income - monthly_expenses
 
-yearly_project_saving = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
+yearly_project_saving = int(monthly_saving * 12 + (monthly_saving * 12 * 0.05))
 
 print(f'Your monthly savings are ${monthly_saving}.')
